+++ docs/source/changelog.rst
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 `Unreleased`_
 -------------
 
+`0.7.0`_ -- 2018-10-15
+----------------------
+
 Changed
 ^^^^^^^
 
@@ -221,7 +224,8 @@ Changed
 * Initial release.
 
 
-.. _Unreleased: https://github.com/Bogdanp/molten/compare/v0.6.1...HEAD
+.. _Unreleased: https://github.com/Bogdanp/molten/compare/v0.7.0...HEAD
+.. _0.7.0: https://github.com/Bogdanp/molten/compare/v0.6.1...v0.7.0
 .. _0.6.1: https://github.com/Bogdanp/molten/compare/v0.6.0...v0.6.1
 .. _0.6.0: https://github.com/Bogdanp/molten/compare/v0.5.2...v0.6.0
 .. _0.5.2: https://github.com/Bogdanp/molten/compare/v0.5.1...v0.5.2